// client/src/types/Drawing.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DrawingType } from "./DrawingType";
import type { JsonValue } from "./serde_json/JsonValue";

export type Drawing = { id: string, layer_id: string, drawing_type: DrawingType, points: JsonValue, stroke_color: string, stroke_width: number, stroke_opacity: number, fill_color: string | null, fill_opacity: number, created_at: string, };